backend/requirements.txt
redis
fastapi
uvicorn[standard]
pymongo
langgraph
langchain
langchain-core
google-genai
pydantic
pydantic-settings
openai
langgraph-checkpoint-mongodb
langgraph-store-mongodb
sentence-transformers
numpy
langchain-openai
langchain-groq
PyPDF2
python-docx
python-multipart
deepgram-sdk
websockets